+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Darkmode style fixes
 - Fix stretched avatars in channel list
 - Fixed file permssions on Linux
+- Reworked file/folder/remote URL fetching dialogs
 
 ## [0.0.2] - 2020-07-31
 ### Added